requirements.txt
aylien-apiclient==0.6.0
beautifulsoup4==4.5.1
bs4==0.0.1
click==6.6
Flask==0.11.1
html5lib==0.999999999
httplib2==0.9.2
isodate==0.5.4
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pyparsing==2.1.10
rdflib==4.2.1
requests==2.11.1
six==1.10.0
SPARQLWrapper==1.6.0
webencodings==0.5
Werkzeug==0.11.11
pyspotlight==0.7.1
goslate==1.5.1
futures==3.0.5
